backend-master/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
pillow==10.4.0
django-cors-headers==4.4.0
django-environ==0.11.2
requests==2.31.0
